shared/requirements-telemetry.txt
# OpenTelemetry Core
opentelemetry-api>=1.20.0,<2.0.0
opentelemetry-sdk>=1.20.0,<2.0.0

# OpenTelemetry Exporters
opentelemetry-exporter-otlp-proto-grpc>=1.20.0,<2.0.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation>=0.41b0,<1.0.0
opentelemetry-instrumentation-fastapi>=0.41b0,<1.0.0
opentelemetry-instrumentation-asyncpg>=0.41b0,<1.0.0
opentelemetry-instrumentation-redis>=0.41b0,<1.0.0
opentelemetry-instrumentation-requests>=0.41b0,<1.0.0
opentelemetry-instrumentation-httpx>=0.41b0,<1.0.0
